Create hopfieldnetwork.c
The Hopfield Network

This program is copyright © 1996 by the author. It is made available as is, and no warranty - about the program, its performance, or its conformity to any specification - is given or implied. It may be used, modified, and distributed freely for private and commercial purposes, as long as the original author is credited as part of the final work.
Hopfield Network Simulator more on : https://courses.cs.washington.edu/courses/cse599/01wi/admin/Assignments/hopfield.html